Add first files via upload
Original files - Predictive model determining the axis of the dominant movement (X or Y) of a 3D printer. 

1000 video files (500 for major X movement, 500 for major Y movement) have been transformed in grayscale and with movement detection, on file 01.
File 02 permits the training of a prediction model based on the transformed videos. 
File 03 permits tranforming some other videos for the test of the prediction model.
File 04 gives the prediction from the transformed files from file 03 and checks reliability of the model.